move to env variable for deciding lang
